Module HelperMethods working
