+++ lib/league_stats.rb
@@ -1,9 +1,9 @@
 require 'CSV'
 require_relative './game_teams'
 require_relative './game'
-require_relative './team'
+require_relative './helper_methods'
 
-class LeagueStats
+class LeagueStats < HelperMethods
 
   attr_reader :game_teams,
               :teams,
@@ -16,89 +16,6 @@ class LeagueStats
     load_game_teams(filepath1)
     load_games(filepath2)
     load_teams(filepath3)
-  end
-
-  def load_game_teams(filepath1)
-    CSV.foreach(filepath1, headers: true, header_converters: :symbol) do |data|
-      @game_teams << GameTeams.new(data)
-    end
-  end
-
-  def load_games(filepath2)
-    CSV.foreach(filepath2, headers: true, header_converters: :symbol) do |data|
-      @games << Game.new(data)
-    end
-  end
-
-  def load_teams(filepath3)
-    CSV.foreach(filepath3, headers: true, header_converters: :symbol) do |data|
-      @teams << Team.new(data)
-    end
-  end
-
-  def game_teams_find_by_game_id(game_id)
-    @game_teams.find do |season_stat|
-      season_stat.game_id == game_id
-    end
-  end
-
-  def games_find_by_game_id(game_id)
-    @games.find do |season_stat|
-      season_stat.game_id == game_id
-    end
-  end
-
-  def teams_find_by_team_id(team_id)
-    @teams.find do |season_stat|
-      season_stat.team_id == team_id
-    end
-  end
-
-  def largest_hash_value(hash)
-    hash.max_by{|k,v| v}
-  end
-
-  def smallest_hash_value(hash)
-    hash.min_by{|k,v| v}
-  end
-
-  def find_teams_by_team_id(game_teams)
-    game_teams.group_by do |game_team|
-      game_team.team_id
-    end
-  end
-
-  def find_away_games(game_teams)
-    game_teams.find_all do |game_team|
-      game_team.hoa == "away"
-    end
-  end
-
-  def find_home_games(game_teams)
-    game_teams.find_all do |game_team|
-      game_team.hoa == "home"
-    end
-  end
-
-  def all_the_goals(teams_by_id)
-    team_and_total_score = {}
-    teams_by_id.each do |team|
-      goals_by_team = team[1].sum do |the_goals|
-        the_goals.goals
-      end
-      team_and_total_score[team[0]] = goals_by_team
-    end
-    team_and_total_score
-  end
-
-  def find_the_teamname(top_or_bottom_scorer)
-    best_or_worst_team = []
-    teams.each do |team|
-      if team.team_id == top_or_bottom_scorer
-        best_or_worst_team << team.teamname
-      end
-    end
-    best_or_worst_team
   end
 
   # #LEAGUE STATS METHODS
@@ -115,7 +32,7 @@ class LeagueStats
       #next find the key that has the highest value, and assign that key to @top_scorer
       top_scorer = largest_hash_value(team_and_total_score)[0]
       #next, find the team name that correlates to the @top_scorer
-      best_team = find_the_teamname(top_scorer)
+      best_team = find_team_name(top_scorer)
       #finally, print the teamname.
       best_team[0]
     end
@@ -124,7 +41,7 @@ class LeagueStats
       teams_by_id = find_teams_by_team_id(game_teams)
       team_and_total_score = all_the_goals(teams_by_id)
       bottom_scorer = smallest_hash_value(team_and_total_score)[0]
-      worst_team = find_the_teamname(bottom_scorer)
+      worst_team = find_team_name(bottom_scorer)
       worst_team[0]
     end
 
@@ -133,7 +50,7 @@ class LeagueStats
       teams_by_id = find_teams_by_team_id(away_games)
       team_and_total_score = all_the_goals(teams_by_id)
       top_scorer = largest_hash_value(team_and_total_score)[0]
-      best_team = find_the_teamname(top_scorer)
+      best_team = find_team_name(top_scorer)
       best_team[0]
     end
 
@@ -142,7 +59,7 @@ class LeagueStats
       teams_by_id = find_teams_by_team_id(home_games)
       team_and_total_score = all_the_goals(teams_by_id)
       top_scorer = largest_hash_value(team_and_total_score)[0]
-      best_team = find_the_teamname(top_scorer)
+      best_team = find_team_name(top_scorer)
       best_team[0]
     end
 
@@ -151,7 +68,7 @@ class LeagueStats
       teams_by_id = find_teams_by_team_id(away_games)
       team_and_total_score = all_the_goals(teams_by_id)
       bottom_scorer = smallest_hash_value(team_and_total_score)[0]
-      worst_team = find_the_teamname(bottom_scorer)
+      worst_team = find_team_name(bottom_scorer)
       worst_team[0]
     end
 
@@ -160,7 +77,7 @@ class LeagueStats
       teams_by_id = find_teams_by_team_id(home_games)
       team_and_total_score = all_the_goals(teams_by_id)
       bottom_scorer = smallest_hash_value(team_and_total_score)[0]
-      worst_team = find_the_teamname(bottom_scorer)
+      worst_team = find_team_name(bottom_scorer)
       worst_team[0]
     end
 end
