+++ lib/season_stats.rb
@@ -2,8 +2,9 @@ require 'CSV'
 require_relative './game_teams'
 require_relative './game'
 require_relative './team'
+require_relative './helper_methods'
 
-class SeasonStats
+class SeasonStats < HelperMethods
 
   attr_reader :game_teams,
               :teams,
@@ -16,181 +17,6 @@ class SeasonStats
     load_game_teams(filepath1)
     load_games(filepath2)
     load_teams(filepath3)
-  end
-
-  def load_game_teams(filepath1)
-    CSV.foreach(filepath1, headers: true, header_converters: :symbol) do |data|
-      @game_teams << GameTeams.new(data)
-    end
-  end
-
-  def load_games(filepath2)
-    CSV.foreach(filepath2, headers: true, header_converters: :symbol) do |data|
-      @games << Game.new(data)
-    end
-  end
-
-  def load_teams(filepath3)
-    CSV.foreach(filepath3, headers: true, header_converters: :symbol) do |data|
-      @teams << Team.new(data)
-    end
-  end
-
-  def game_teams_find_by_game_id(game_id)
-    @game_teams.find do |season_stat|
-      season_stat.game_id == game_id
-    end
-  end
-
-  def games_find_by_game_id(game_id)
-    @games.find do |season_stat|
-      season_stat.game_id == game_id
-    end
-  end
-
-  def teams_find_by_team_id(team_id)
-    @teams.find do |season_stat|
-      season_stat.team_id == team_id
-    end
-  end
-
-  def largest_hash_value(hash)
-    hash.max_by{|k,v| v}
-  end
-
-  def smallest_hash_value(hash)
-    hash.min_by{|k,v| v}
-  end
-
-  def find_win_games(game_teams)
-    game_teams.find_all do |game_team|
-      game_team.result == "WIN"
-    end
-  end
-
-  def find_lose_games(game_teams)
-    game_teams.find_all do |game_team|
-      game_team.result == "LOSS"
-    end
-  end
-
-  def find_result_games_with_key_as_game_id(result_games)
-    result_games.group_by do |game_won|
-      game_won.game_id
-    end
-  end
-
-  def find_this_season(the_season)
-    this_season = []
-    games.find_all do |game_in_season|
-      if game_in_season.season == the_season
-        this_season << game_in_season
-      end
-    end
-    this_season
-  end
-
-  def find_result_games_this_season(this_season, result_games_with_key_as_game_id)
-    result_games_this_season = []
-    this_season.each do |the_game|
-      if result_games_with_key_as_game_id.keys.any?(the_game.game_id) == true
-        result_games_this_season << the_game
-      end
-    end
-    result_games_this_season
-  end
-
-  def find_games_by_game_id(games_this_season)
-    result_games_by_game_id = []
-    games_this_season.group_by do |game_this_season|
-      result_games_by_game_id << game_this_season.game_id
-    end
-    result_games_by_game_id
-  end
-
-  def find_game_list(games_with_key_as_game_id, games_by_game_id)
-    game_list = []
-    games_with_key_as_game_id.find_all do |game_result|
-      if games_by_game_id.any?(game_result[0]) == true
-        game_list << game_result[1]
-      end
-    end
-    game_list
-  end
-
-  def find_game_list_with_reduce(games_with_key_as_game_id, games_by_game_id)
-    game_list = []
-    games_with_key_as_game_id.find_all do |game_result|
-      if games_by_game_id.any?(game_result[0]) == true
-        game_list << game_result[1].reduce
-      end
-    end
-    game_list
-  end
-
-  def find_teams_by_team_id(game_teams)
-    game_teams.group_by do |game_team|
-      game_team.team_id
-    end
-  end
-
-  def find_teams_by_game_id(game_teams)
-    game_teams.group_by do |game_team|
-      game_team.game_id
-    end
-  end
-
-  def find_team_and_results(teams_by_id, this_season)
-    find_team_and_results = {}
-    teams_by_id.each do |team|
-      find_team_and_results[team[0]] = team[1].count.to_f / this_season.count.to_f
-    end
-    find_team_and_results
-  end
-
-  def find_team_and_accuracy(teams_by_id)
-    team_and_accuracy = {}
-    teams_by_id.each do |team|
-      goals_by_team = team[1].sum do |the_goals|
-        the_goals.goals.to_f
-      end
-      shots_by_team = team[1].sum do |the_shots|
-        the_shots.shots.to_f
-      end
-      team_and_accuracy[team[0]] = goals_by_team / shots_by_team
-    end
-    team_and_accuracy
-  end
-
-  def find_team_and_tackles(teams_by_id)
-    team_and_total_tackles = {}
-    teams_by_id.each do |team|
-      goals_by_team = team[1].sum do |the_tackles|
-        the_tackles.tackles
-      end
-      team_and_total_tackles[team[0]] = goals_by_team
-    end
-    team_and_total_tackles
-  end
-
-  def find_coach_name(best_or_worst_coach)
-    coach_name = []
-    game_teams.each do |team|
-      if team.team_id == best_or_worst_coach
-        coach_name << team.head_coach
-      end
-    end
-    coach_name
-  end
-
-  def find_team_name(best_or_worst_team)
-    team_name = []
-    teams.each do |team|
-      if team.team_id == best_or_worst_team
-        team_name << team.teamname
-      end
-    end
-    team_name
   end
 
 # Name of the Coach with the best win percentage for the season
